Add authentication status to JWT and make modular functions for resetting state on various actions. Apply these functions to the various actions such as leave trip, sign out, delete trip.

+++ backend/src/routers/TripRouter.py
@@ -4,8 +4,9 @@ from models.DatabaseHandler import db_handler
 from models.Schemas import Trip, Traveller, Itinerary, Message, NewPollBody, PollResponseBody, Packing
 from typing import Annotated
 from datetime import date, datetime
-from models.S3Handler import minio_client
 from utilities.merge_polls import merge_polls
+from utilities.auth_helpers import verify_attendance, verify_admin
+from models.S3Handler import minio_client
 
 
 trip_router = APIRouter(
@@ -20,7 +21,7 @@ async def create_trip(
     description: Annotated[str, Form()],
     start_date: Annotated[date, Form()],
     end_date: Annotated[date, Form()]
-    ) -> Trip | dict[str, str]:   
+    ) -> Trip | dict[str, str]:
 
     try:
         # This call must not only create the trip, but must add
@@ -32,7 +33,7 @@ async def create_trip(
                 (destination, description, start_date, end_date, created_by)
                 VALUES (%s, %s, %s, %s, %s)
                 RETURNING id
-            ) INSERT INTO traveller_trip VALUES (%s, (SELECT id from temp_table)) RETURNING trip_id;
+            ) INSERT INTO traveller_trip VALUES (%s, (SELECT id from temp_table), TRUE, TRUE) RETURNING trip_id;
                                 
         """, (destination, description, start_date, end_date, request.state.user['email'], request.state.user['id']))
         trip_id = data[0]['trip_id']
@@ -56,6 +57,8 @@ async def create_trip(
 @trip_router.delete('/{trip_id}')
 async def delete_trip(request: Request, trip_id: str) -> dict[str, str]:
     try:
+        verify_admin(trip_id, request.state.user['trips'])
+
         db_handler.query("""
             DELETE FROM trip WHERE id = %s;
         """, (trip_id,))
@@ -79,8 +82,10 @@ async def delete_trip(request: Request, trip_id: str) -> dict[str, str]:
 
 # Get itinerary entries for this trip
 @trip_router.get('/{trip_id}/itinerary')
-async def get_itinerary_info(trip_id: str) -> list[Itinerary] | dict[str, str]:
+async def get_itinerary_info(request: Request, trip_id: str) -> list[Itinerary] | dict[str, str]:
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+
         itinerary = db_handler.query("""
             SELECT * FROM itinerary WHERE trip_id = %s ORDER BY start_time;
         """, (trip_id,))
@@ -109,6 +114,8 @@ async def add_itinerary_stop(
     ) ->  dict[str, str]:
 
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+
         db_handler.query("""
         INSERT INTO itinerary (trip_id, title, description, start_time, end_time, created_by)
         VALUES (%s, %s, %s, %s, %s, %s);
@@ -131,10 +138,12 @@ async def add_itinerary_stop(
         )
 
 
-# Add itinerary item for this trip
+# Remove itinerary item from this trip
 @trip_router.delete('/{trip_id}/itinerary/{item_id}')
-async def remove_itinerary_stop( trip_id: str, item_id: int) ->  dict[str, str]:
+async def remove_itinerary_stop(request: Request, trip_id: str, item_id: int) ->  dict[str, str]:
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+
         # Even though we technically only need the item id to delete, we
         # will also use the trip id for improved robustness since we have it
         db_handler.query("""
@@ -158,8 +167,10 @@ async def remove_itinerary_stop( trip_id: str, item_id: int) ->  dict[str, str]:
         )
 
 @trip_router.get('/{trip_id}/message')
-async def get_messages(trip_id: str) -> list[Message] | dict[str, str]:
+async def get_messages(request: Request, trip_id: str) -> list[Message] | dict[str, str]:
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+
         data = db_handler.query("""
             SELECT * FROM message WHERE trip_id = %s ORDER BY id;
         """, (trip_id,))
@@ -177,8 +188,10 @@ async def get_messages(trip_id: str) -> list[Message] | dict[str, str]:
 
     
 @trip_router.get('/{trip_id}/poll')
-async def get_polls(trip_id: str) -> list[PollResponseBody] | dict[str, str]:
+async def get_polls(request: Request, trip_id: str) -> list[PollResponseBody] | dict[str, str]:
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+
         # The results for one poll on this trip will have the following number
         # of rows:
         # M options * N people who voted this option, M >= 1 & N >= 1
@@ -228,6 +241,8 @@ async def get_polls(trip_id: str) -> list[PollResponseBody] | dict[str, str]:
 @trip_router.post('/{trip_id}/poll')
 async def add_poll(request: Request, trip_id: str, poll_body: NewPollBody) -> dict[str, str]:
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+
         res = db_handler.query("""
             INSERT INTO poll 
                 (trip_id, title, anonymous, description, created_by) 
@@ -263,12 +278,12 @@ async def add_poll(request: Request, trip_id: str, poll_body: NewPollBody) -> di
 @trip_router.delete('/{trip_id}/poll/{poll_id}')
 async def delete_poll(request: Request, trip_id: str, poll_id: int) -> dict[str, str]:
     try:
-        # Only allow deletion of a poll by the creator, or allow anyone to delete
-        # if the creator deleted account (creator is null)
+        verify_admin(trip_id, request.state.user['trips'])
+
         # Even though we technically don't need the trip_id to delete, we
         # will also use it for improved robustness since we have it
         db_handler.query("""
-            DELETE FROM poll WHERE trip_id=%s AND id=%s AND (created_by=%s OR created_by IS NULL);
+            DELETE FROM poll WHERE trip_id=%s AND id=%s;
         """, (trip_id, poll_id, request.state.user['email']))
 
         return JSONResponse(
@@ -289,8 +304,10 @@ async def delete_poll(request: Request, trip_id: str, poll_id: int) -> dict[str,
 
 
 @trip_router.get('/{trip_id}/packing')
-async def get_packing_items(trip_id: str) -> list[Packing] | dict[str, str]:
+async def get_packing_items(request: Request, trip_id: str) -> list[Packing] | dict[str, str]:
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+
         data = db_handler.query("""
             SELECT * FROM packing WHERE trip_id=%s ORDER BY id;
         """, (trip_id,))
@@ -316,6 +333,8 @@ async def add_packing_item(
     ) -> dict[str, str]:
 
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+
         db_handler.query("""
             INSERT INTO packing (trip_id, item, quantity, description, created_by) 
             VALUES (%s, %s, %s, %s, %s)
@@ -339,8 +358,10 @@ async def add_packing_item(
 
 
 @trip_router.delete('/{trip_id}/packing/{item_id}')
-async def delete_packing_item(trip_id: str, item_id: int) -> dict[str, str]:
+async def delete_packing_item(request: Request, trip_id: str, item_id: int) -> dict[str, str]:
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+
         # Even though we technically only need the item id to delete, we
         # will also use the trip id for improved robustness since we have it
         db_handler.query("""
@@ -367,6 +388,8 @@ async def delete_packing_item(trip_id: str, item_id: int) -> dict[str, str]:
 @trip_router.patch('/{trip_id}/packing/claim/{item_id}')
 async def claim_packing_item(request: Request, trip_id: str, item_id: int) -> dict[str, str]:
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+
         # Allow a user to claim this item as long as it is not currently claimed
         db_handler.query("""
             UPDATE packing SET packed_by = %s WHERE packed_by IS NULL AND id = %s;
@@ -390,8 +413,10 @@ async def claim_packing_item(request: Request, trip_id: str, item_id: int) -> di
 
 
 @trip_router.patch('/{trip_id}/packing/unclaim/{item_id}')
-async def unclaim_packing_item(trip_id: str, item_id: int) -> dict[str, str]:
+async def unclaim_packing_item(request: Request, trip_id: str, item_id: int) -> dict[str, str]:
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+
         db_handler.query("""
             UPDATE packing SET packed_by = NULL WHERE packed_by IS NOT NULL AND id = %s;
         """, (item_id,))
@@ -415,8 +440,10 @@ async def unclaim_packing_item(trip_id: str, item_id: int) -> dict[str, str]:
 
 # Get info for travellers on this trip
 @trip_router.get('/{trip_id}/travellers')
-async def get_travellers(trip_id: str) ->  list[Traveller] | dict[str, str]:
+async def get_travellers(request: Request, trip_id: str) ->  list[Traveller] | dict[str, str]:
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+        
         travellers = db_handler.query("""
             SELECT * from traveller WHERE id in (SELECT traveller_id FROM traveller_trip WHERE trip_id=%s) ORDER BY last_name;
             """, (trip_id,))
@@ -461,6 +488,8 @@ async def request_join_trip(request: Request, trip_id: str) -> dict[str, str]:
 @trip_router.patch('/{trip_id}/travellers/accept/{requestor_id}')
 async def accept_join_trip(request: Request, trip_id: str, requestor_id: str) -> dict[str, str]:
     try:
+        verify_admin(trip_id, request.state.user['trips'])
+
         db_handler.query("""
             UPDATE traveller_trip SET confirmed=TRUE WHERE traveller_id=%s AND trip_id=%s;
         """, (requestor_id, trip_id))
@@ -486,6 +515,8 @@ async def accept_join_trip(request: Request, trip_id: str, requestor_id: str) ->
 @trip_router.delete('/{trip_id}/travellers/remove/{requestor_id}')
 async def deny_join_trip(request: Request, trip_id: str, requestor_id: str) -> dict[str, str]:
     try:
+        verify_admin(trip_id, request.state.user['trips'])
+
         db_handler.query("""
             DELETE FROM traveller_trip WHERE traveller_id=%s AND trip_id=%s;
         """, (requestor_id, trip_id))
@@ -502,15 +533,17 @@ async def deny_join_trip(request: Request, trip_id: str, requestor_id: str) -> d
         return JSONResponse(
             status_code=500,
             content = {
-                "message": f"ERROR: Unable to reject user {requestor_id} from trip {trip_id}"
+                "message": f"ERROR: Error while submitting rejection for user {requestor_id} from trip {trip_id}"
             }
         )
 
 
 # User leaves a trip
-@trip_router.post('/{trip_id}/travellers/leave')
+@trip_router.delete('/{trip_id}/travellers/leave')
 async def leave_trip(request: Request, trip_id: str) -> dict[str, str]:
     try:
+        verify_attendance(trip_id, request.state.user['trips'])
+
         db_handler.query("""
             DELETE FROM traveller_trip WHERE traveller_id=%s AND trip_id=%s;
         """, (request.state.user['id'], trip_id))
